refactor: Use same type for db and app

+++ src/main/resources/schema.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS products (
-    id INT PRIMARY KEY AUTO_INCREMENT,
+    id BIGINT PRIMARY KEY AUTO_INCREMENT,
     product_name VARCHAR(255) NOT NULL,
     price DOUBLE CHECK (price >= 0),
     image_url VARCHAR(255)
-    );
+);
